Revert "New functionality (#11)"
This reverts commit f860385bdcb100aaead877088a4fd3bfee43a7f3.
