--- src/main/java/pl/kurs/figures/exceptions/EmptyTableException.java
@@ -1,12 +1,0 @@
-package pl.kurs.figures.exceptions;
-
-public class EmptyTableException extends RuntimeException {
-
-    public EmptyTableException(String message) {
-        super(message);
-    }
-
-    public EmptyTableException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
