--- src/main/java/pl/kurs/figures/security/dto/UserDTO.java
@@ -1,19 +1,0 @@
-package pl.kurs.figures.security.dto;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import pl.kurs.figures.security.model.Role;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class UserDTO {
-    private Long id;
-    private String username;
-    private Role role;
-    private Integer figuresNumber;
-}
