+++ src/main/resources/db/changelog/2024/02/08-01-changelog.xml
@@ -145,10 +145,4 @@
         </createView>
     </changeSet>
 
-    <changeSet id="1707403054540-3" author="mateuszparasiewicz">
-        <addColumn tableName="users">
-            <column name="figures_number" type="INT"></column>
-        </addColumn>
-    </changeSet>
-
 </databaseChangeLog>
