+++ src/test/java/pl/kurs/figures/controller/FigureControllerTest.java
@@ -59,6 +59,7 @@ class FigureControllerTest {
                 .build();
     }
 
+
     public User setupUser() {
         BCryptPasswordEncoder bCryptPasswordEncoder = new BCryptPasswordEncoder();
         Role userRole = Role.USER;
