--- src/test/java/pl/kurs/figures/security/controller/UserControllerTest.java
@@ -1,57 +1,0 @@
-package pl.kurs.figures.security.controller;
-
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.http.MediaType;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import org.springframework.web.context.WebApplicationContext;
-import pl.kurs.figures.security.model.Role;
-import pl.kurs.figures.security.model.User;
-import pl.kurs.figures.security.repository.UserRepository;
-
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-
-@SpringBootTest
-@AutoConfigureMockMvc
-class UserControllerTest {
-
-    @Autowired
-    private MockMvc mockMvc;
-    @Autowired
-    private UserRepository userRepository;
-    @Autowired
-    private WebApplicationContext webApplicationContext;
-    @BeforeEach
-    public void setup() {
-        this.mockMvc = MockMvcBuilders
-                .webAppContextSetup(this.webApplicationContext)
-                .build();
-    }
-
-    public User setupUser() {
-        BCryptPasswordEncoder bCryptPasswordEncoder = new BCryptPasswordEncoder();
-        Role userRole = Role.USER;
-        User testUser = new User("user", bCryptPasswordEncoder.encode("user"), userRole);
-        return userRepository.save(testUser);
-    }
-    @Test
-    void shouldReturnUsersPage() throws Exception{
-        mockMvc.perform(get("/api/v1/users")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .param("page", "0")
-                        .param("size", "10"))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.content").exists())
-                .andExpect(jsonPath("$.content").isArray())
-                .andExpect(jsonPath("$.content[0].username").isString())
-                .andExpect(jsonPath("$.content[0].role").isString());
-    }
-}
