--- src/test/java/pl/kurs/figures/security/service/UserServiceImplTest.java
@@ -1,10 +1,0 @@
-package pl.kurs.figures.security.service;
-
-import org.junit.jupiter.api.Test;
-
-class UserServiceImplTest {
-
-    @Test
-    void getAll() {
-    }
-}
